Moved all the goofy swapping about and stuff in Anim blend down into the subanims where it is a bit more reasonable.  Gets rid of a const problem as well.
Constify the foreign skeleton in the checker struct

Use int64 for the name to index dictionary.  This is annoying but it
gets rid of warnings.

Warning fixes all over

Removed the const on the vert data passed into Mesh.  I might revisit
this and try to keep it const in the structure because it shouldn't ever
be modified, just read.

Take out the stuffkeeper ref in ParticleBoss so it can be kept const

Logic fix in wrapping menu selection code

+++ MeshLib/Skeleton.c
@@ -161,7 +161,7 @@ void	Skeleton_FillBoneArray(const Skeleton *pSkel,
 static void	srMakeNameDict(Skeleton *pSkel, GSNode *pNode)
 {
 	printf("srMakeNameDict: Adding to name dictionary %s index %d\n", utstring_body(pNode->szName), pNode->mIndex);
-	DictSZ_Add(&pSkel->mpNameToIndex, pNode->szName, (void *)pNode->mIndex);
+	DictSZ_Add(&pSkel->mpNameToIndex, pNode->szName, (void *)((int64_t)pNode->mIndex));
 
 	for(int i=0;i < pNode->mNumChildren;i++)
 	{
